// frontend/src/openapi/generated/model/memo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zete
 * 메모 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Tag } from './tag';

/**
 * 
 * @export
 * @interface Memo
 */
export interface Memo {
    /**
     * 
     * @type {number}
     * @memberof Memo
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof Memo
     */
    'createAt': string;
    /**
     * 
     * @type {string}
     * @memberof Memo
     */
    'updateAt': string;
    /**
     * 
     * @type {string}
     * @memberof Memo
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Memo
     */
    'content'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Memo
     */
    'isImportant'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof Memo
     */
    'userId'?: number;
    /**
     * 
     * @type {number}
     * @memberof Memo
     */
    'cateId'?: number;
    /**
     * 
     * @type {Array<Tag>}
     * @memberof Memo
     */
    'tags'?: Array<Tag>;
}